Migrate tears of guthix to slash command
